Move database and relative component to own dir
