+++ lib/database/index.js
@@ -6,7 +6,7 @@
 
 'use strict';
 
-const paths = require('./paths');
+const paths = require('../paths');
 const store = require('./store');
 const ruleset = require('./ruleset');
 const results = require('./results');
