+++ lib/database/results.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const pad = require('./pad');
+const pad = require('../pad');
 
 /**
  * Hold an evaluation result.
