+++ lib/database/ruleset.js
@@ -4,8 +4,8 @@
 
 'use strict';
 
-const parser = require('./parser');
-const paths = require('./paths');
+const parser = require('../parser');
+const paths = require('../paths');
 const results = require('./results');
 
 /**
