+++ lib/database/store.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const paths = require('./paths');
+const paths = require('../paths');
 const primitiveTypes = new Set(['string', 'number', 'boolean']);
 const priorityTypes = new Set(['string', 'number']);
 const invalidChar = ['.', '$', '[', ']', '#', '/'];
