+++ lib/util.js
@@ -4,8 +4,8 @@
 
 'use strict';
 
-const ruleset = require('./ruleset');
-const store = require('./store');
+const ruleset = require('./database/ruleset');
+const store = require('./database/store');
 const database = require('./database');
 const pad = require('./pad');
